requirements.txt
streamlit==1.49.1
google-generativeai==0.8.5